abandoned WIP: move nexus-external-endpoints into a crate
