+++ Cargo.toml
@@ -42,6 +42,7 @@ members = [
     "nexus/db-model",
     "nexus/db-queries",
     "nexus/defaults",
+    "nexus/external-endpoints",
     "nexus/inventory",
     "nexus/macros-common",
     "nexus/reconfigurator/execution",
@@ -121,6 +122,7 @@ default-members = [
     "nexus/db-model",
     "nexus/db-queries",
     "nexus/defaults",
+    "nexus/external-endpoints",
     "nexus/inventory",
     "nexus/reconfigurator/execution",
     "nexus/reconfigurator/planning",
@@ -260,6 +262,7 @@ nexus-config = { path = "nexus-config" }
 nexus-db-model = { path = "nexus/db-model" }
 nexus-db-queries = { path = "nexus/db-queries" }
 nexus-defaults = { path = "nexus/defaults" }
+nexus-external-endpoints = { path = "nexus/external-endpoints" }
 nexus-inventory = { path = "nexus/inventory" }
 nexus-macros-common = { path = "nexus/macros-common" }
 nexus-reconfigurator-execution = { path = "nexus/reconfigurator/execution" }
