+++ nexus/Cargo.toml
@@ -81,6 +81,7 @@ uuid.workspace = true
 nexus-defaults.workspace = true
 nexus-db-model.workspace = true
 nexus-db-queries.workspace = true
+nexus-external-endpoints.workspace = true
 nexus-inventory.workspace = true
 nexus-reconfigurator-execution.workspace = true
 nexus-reconfigurator-planning.workspace = true
