+++ nexus/external-endpoints/Cargo.toml
@@ -1,0 +1,39 @@
+[package]
+name = "nexus-external-endpoints"
+version = "0.1.0"
+edition = "2021"
+
+[build-dependencies]
+omicron-rpaths.workspace = true
+
+[dependencies]
+anyhow.workspace = true
+dropshot.workspace = true
+hex.workspace = true
+http.workspace = true
+hyper.workspace = true
+nexus-db-model.workspace = true
+nexus-db-queries.workspace = true
+nexus-types.workspace = true
+omicron-common.workspace = true
+openssl.workspace = true
+# See omicron-rpaths for more about the "pq-sys" dependency.
+pq-sys = "*"
+rcgen.workspace = true
+reqwest.workspace = true
+rustls.workspace = true
+rustls-pemfile.workspace = true
+serde_with.workspace = true
+serde.workspace = true
+slog.workspace = true
+thiserror.workspace = true
+tokio.workspace = true
+uuid.workspace = true
+
+omicron-workspace-hack.workspace = true
+
+[dev-dependencies]
+chrono.workspace = true
+http.workspace = true
+omicron-test-utils.workspace = true
+schemars.workspace = true
