+++ nexus/reconfigurator/execution/Cargo.toml
@@ -15,6 +15,7 @@ internal-dns.workspace = true
 nexus-config.workspace = true
 nexus-db-model.workspace = true
 nexus-db-queries.workspace = true
+nexus-external-endpoints.workspace = true
 nexus-types.workspace = true
 omicron-common.workspace = true
 reqwest.workspace = true
