+++ nexus/reconfigurator/execution/src/dns.rs
@@ -13,6 +13,7 @@ use nexus_db_model::Silo;
 use nexus_db_queries::context::OpContext;
 use nexus_db_queries::db::datastore::DnsVersionUpdateBuilder;
 use nexus_db_queries::db::DataStore;
+use nexus_external_endpoints::ExternalEndpoints;
 use nexus_types::deployment::Blueprint;
 use nexus_types::deployment::OmicronZoneType;
 use nexus_types::internal_api::params::DnsConfigParams;
@@ -62,9 +63,11 @@ pub(crate) async fn deploy_dns(
     // Next, construct the DNS config represented by the blueprint.
     let internal_dns_config_blueprint =
         blueprint_internal_dns_config(blueprint, sleds_by_id);
-    let silos = todo!(); // XXX-dap
+    let external_endpoints = read_all_endpoints(datastore, opctx)
+        .await
+        .internal_context("reading external endpoints to deploy DNS")?;
     let external_dns_config_blueprint =
-        blueprint_external_dns_config(blueprint, silos);
+        blueprint_external_dns_config(blueprint, &external_endpoints);
 
     // Deploy the changes.
     deploy_dns_one(
@@ -302,8 +305,15 @@ pub fn blueprint_internal_dns_config(
 
 pub fn blueprint_external_dns_config(
     blueprint: &Blueprint,
-    silos: Vec<Silo>,
+    endpoints: &ExternalEndpoints,
 ) -> DnsConfigParams {
+    let dns_params = DnsConfigParams {
+        generation: blueprint.external-dns_version.next(),
+        time_created: Utc::now(),
+        zones: vec![DnsConfigZone {
+
+        }]
+    };
     todo!(); // XXX-dap
 }
